Introduce saving user config to flash memory
Now, device is saving user config to non volatile memory.
We are reserving space for user config in the device's
flash memory.

The main purpose of this utility, is to handle unexpected
loss of electrical power.
We want to let the device to boot itself up with the previous config.